requirements.txt
# AI Quant Research Hub (AIQRH) - MVP Dependencies

# Core
python>=3.10
pyyaml>=6.0

# Data
pandas>=2.0.0
numpy>=1.24.0
akshare>=1.12.0
requests>=2.28.0

# Database
pymongo>=4.5.0
sqlalchemy>=2.0.0

# Analysis & Risk
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
plotly>=5.18.0
matplotlib>=3.7.0

# Report
weasyprint>=60.0
jinja2>=3.1.0

# Scheduling
apscheduler>=3.10.0

# Utils
python-dotenv>=1.0.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
motor>=3.3.0
